changed BrowserGuesser to abstract class
